Time: 193 ms (25.53%), Space: 53.7 MB (77.22%) - LeetHub

+++ 1642-furthest-building-you-can-reach/1642-furthest-building-you-can-reach.cpp
@@ -4,7 +4,7 @@ public:
         
         int n=h.size();
         
-        priority_queue<long long,vector<long long>,greater<long long>> q;
+        priority_queue<int,vector<int>,greater<int>> q;
         
         long long s=0,ans=-1;
